refactor(admin/projects): ProjectFormDrawer 디렉터리 분리 + constants/import 정리
- ProjectFormDrawer 단일 파일을 디렉터리 모듈로 분리 (index.tsx + components/MemberRow + components/ThumbnailUploader)
- 페이지 로컬 constants.ts 제거하고 entities/project 의 PLATFORM_LABEL/PLATFORM_OPTIONS 를 사용
- ProjectsTable / ProjectsToolbar 의 import 경로를 entities/project 로 업데이트
